requirements_cuda114.txt
# AuraNet Requirements for CUDA 11.4
# Compatible with CUDA 11.4 and cuDNN 8.2+

# Core PyTorch stack (CUDA 11.4 compatible)
# Use PyTorch 2.0.1 which is the last stable version with good CUDA 11.4 support
torch==2.0.1+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.15.2+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
torchaudio==2.0.2+cu118 --extra-index-url https://download.pytorch.org/whl/cu118

# Core scientific computing
numpy>=1.21.0,<1.25.0  # Compatible with PyTorch 2.0.1
scipy>=1.7.0,<1.11.0

# Computer vision and image processing
opencv-python>=4.7.0,<4.9.0
Pillow>=9.0.0,<10.0.0
scikit-image>=0.19.0,<0.22.0
imageio>=2.22.0,<2.32.0

# Machine learning utilities
scikit-learn>=1.1.0,<1.4.0
pytorch-metric-learning>=1.6.0,<2.1.0

# Deep learning utilities
timm>=0.6.0,<0.10.0  # Use slightly older timm for better compatibility
einops>=0.6.0,<0.8.0
torchmetrics>=0.11.0,<1.3.0

# Signal processing (for wavelets)
pywavelets>=1.4.0,<1.5.0

# Logging and monitoring
tensorboard>=2.10.0,<2.15.0
tensorboardX>=2.6,<2.7.0
wandb>=0.15.0,<0.17.0

# Configuration and utilities
PyYAML>=6.0,<6.1.0
omegaconf>=2.3.0,<2.4.0
hydra-core>=1.3.0,<1.4.0

# Progress bars and CLI
tqdm>=4.64.0,<4.67.0
click>=8.1.0,<8.2.0
rich>=13.0.0,<14.0.0

# Visualization
matplotlib>=3.6.0,<3.8.0
seaborn>=0.12.0,<0.13.0
plotly>=5.15.0,<5.18.0

# Data handling
pandas>=1.5.0,<2.1.0
h5py>=3.7.0,<3.10.0

# Distributed training utilities
accelerate>=0.20.0,<0.25.0

# Image augmentation
albumentations>=1.3.0,<1.4.0

# Metrics and evaluation
pytorch-lightning>=2.0.0,<2.2.0  # Use compatible version with PyTorch 2.0.1
torcheval>=0.0.6,<0.0.8

# Memory profiling and optimization
psutil>=5.9.0,<5.10.0
py3nvml>=0.2.7,<0.3.0

# SSIM loss implementation
pytorch-ssim>=0.1.0

# Additional contrastive learning utilities (optional - may not be compatible)
# lightly>=1.4.0,<1.5.0  # Comment out if installation fails

# Face detection (if needed for preprocessing)
facenet-pytorch>=2.5.0,<2.6.0

# File I/O optimization
zarr>=2.12.0,<2.17.0

# Development and testing
pytest>=7.0.0,<7.5.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<6.2.0
mypy>=1.0.0,<1.8.0

# Jupyter notebook support (optional but useful)
jupyter>=1.0.0
ipykernel>=6.0.0,<6.30.0
ipywidgets>=8.0.0,<8.2.0

# Memory efficient attention (CUDA 11.4 compatible version)
# xformers==0.0.16  # Specific version for CUDA 11.4 - uncomment if needed

# Gradient accumulation and mixed precision utilities
# apex  # Build from source for CUDA 11.4 compatibility

# Additional CUDA 11.4 specific packages
# cupy-cuda114>=11.0.0  # CuPy for CUDA 11.4 if needed for custom operations

# Alternative SSIM implementation if pytorch-ssim fails
pytorch-msssim>=0.2.1

# Specific versions tested with CUDA 11.4
setuptools>=65.0.0
wheel>=0.38.0
